# examples/showcase/generated/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "zever API",
    "version": "0.0.0"
  },
  "paths": {
    "/orders": {
      "get": {
        "operationId": "ShopService_ListOrders",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/shop_Order"
                      }
                    },
                    "next_cursor": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/orders/checkout": {
      "post": {
        "operationId": "ShopService_Checkout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/shop_CheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/shop_OrderReceipt"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_ARGUMENT: bad coupon; FAILED_PRECONDITION"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "ShopService_GetOrder",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/shop_Order"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/products": {
      "get": {
        "operationId": "ShopService_ListProducts",
        "parameters": [
          {
            "in": "query",
            "name": "category_id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/shop_Product"
                      }
                    },
                    "next_cursor": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "operationId": "ShopService_CreateProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/shop_CreateProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/shop_Product"
                }
              }
            }
          },
          "409": {
            "description": "ALREADY_EXISTS: product sku taken"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "admin"
        ]
      }
    },
    "/products/{id}": {
      "delete": {
        "operationId": "ShopService_DeleteProduct",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/shop_Product"
                }
              }
            }
          },
          "403": {
            "description": "PERMISSION_DENIED: no access"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "admin"
        ],
        "x-permission": {
          "check": "product.delete",
          "resource": "Product",
          "owner_field": "category_id"
        }
      },
      "get": {
        "operationId": "ShopService_GetProduct",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/shop_Product"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "ShopService_PatchProduct",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/shop_PatchProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/shop_Product"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_ARGUMENT: negative stock"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "admin"
        ]
      },
      "put": {
        "operationId": "ShopService_UpdateProduct",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/shop_UpdateProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/shop_Product"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "admin"
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "shop_Category": {
        "type": "object",
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          }
        },
        "required": [
          "id"
        ]
      },
      "shop_CheckoutRequest": {
        "type": "object",
        "properties": {
          "req": {
            "$ref": "#/components/schemas/shop_CheckoutRequest"
          }
        }
      },
      "shop_CreateOrderRequest": {
        "type": "object",
        "properties": {
          "coupon": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "total_cents": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "shop_CreateProductRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "price_cents": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "shop_Order": {
        "type": "object",
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "high"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/shop_OrderStatus"
          },
          "total_cents": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "id"
        ]
      },
      "shop_OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "order_id": {
            "type": "string",
            "format": "uuid"
          },
          "price_cents": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "product_id": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          }
        },
        "required": [
          "id"
        ]
      },
      "shop_OrderReceipt": {
        "type": "object",
        "properties": {
          "order_id": {
            "type": "string",
            "format": "uuid"
          },
          "total_cents": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "shop_OrderStatus": {
        "type": "string",
        "enum": [
          "pending",
          "paid",
          "shipped",
          "cancelled"
        ]
      },
      "shop_PatchProductRequest": {
        "type": "object",
        "properties": {
          "stock": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "shop_Product": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "featured": {
            "type": "boolean"
          },
          "headline": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "price_cents": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "sku": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "stock": {
            "type": "integer",
            "format": "int64"
          },
          "weight": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "id"
        ]
      },
      "shop_Profile": {
        "type": "object",
        "properties": {
          "bio": {
            "type": "string",
            "maxLength": 500
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "id"
        ]
      },
      "shop_Review": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "product_id": {
            "type": "string",
            "format": "uuid"
          },
          "rating": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "maximum": 5
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "id"
        ]
      },
      "shop_Role": {
        "type": "string",
        "enum": [
          "admin",
          "member"
        ]
      },
      "shop_Tag": {
        "type": "object",
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 30
          }
        },
        "required": [
          "id"
        ]
      },
      "shop_UpdateProductRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        }
      },
      "shop_User": {
        "type": "object",
        "properties": {
          "age": {
            "type": "integer",
            "format": "int32",
            "minimum": 0,
            "maximum": 150
          },
          "avatar": {
            "type": "string",
            "format": "byte"
          },
          "birthday": {
            "type": "string",
            "format": "date"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "credit_cents": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "nickname": {
            "type": "string",
            "nullable": true
          },
          "password_hash": {
            "type": "string"
          },
          "prefs": {
            "type": "object"
          },
          "rating": {
            "type": "number",
            "format": "float"
          },
          "role": {
            "$ref": "#/components/schemas/shop_Role"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "verified": {
            "type": "boolean"
          }
        },
        "required": [
          "id"
        ]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}